+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { Analytics } from '@vercel/analytics/react';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
-import { ProductsProvider } from '@/context/ProductsContext';
+import { ProductsProvider } from '@/src/data-access-layer/ProductsContext';
 
 const inter = Inter({ subsets: ['latin'] });
