+++ app/search/page.tsx
@@ -2,7 +2,7 @@
 
 import Header from '@/src/components/layout/Header';
 import ProductGrid from '@/src/components/product/ProductGrid';
-import { useProductsContext } from '@/context/ProductsContext';
+import { useProductsContext } from '@/src/data-access-layer/ProductsContext';
 import { Suspense } from 'react';
 
 export default function Search() {
