+++ src/components/layout/Header.tsx
@@ -1,6 +1,6 @@
 import { Disclosure } from '@headlessui/react';
 import { usePathname } from 'next/navigation';
-import Search from '../search/Search';
+import Search from '../../features/search/Search';
 import Title from './Title';
 import { classNames } from '@/src/utils/CommonFunctions';
