+++ src/features/search/Search.tsx
@@ -2,13 +2,13 @@
 
 import { useCallback, useEffect, useRef, useState } from 'react';
 import ShowSearchResults from './ShowSearchResults';
-import ProductModal from '../product/ProductModal';
+import ProductModal from '../../components/product/ProductModal';
 import SearchBar from './SearchBar';
 import debounce from 'lodash.debounce';
-import { getKassalappProducts } from '@/api/KassalappApi';
-import { getProcessingInfo } from '@/api/ProductEvaluator';
+import { getKassalappProducts } from '@/src/data-access-layer/api/KassalappApi';
+import { getProcessingInfo } from '@/src/data-access-layer/api/ProductEvaluator';
 import { useRouter, useSearchParams } from 'next/navigation';
-import { useProductsContext } from '@/context/ProductsContext';
+import { useProductsContext } from '@/src/data-access-layer/ProductsContext';
 import { Product } from '@/src/types/ProductTypes';
 
 export default function Search(): JSX.Element {
